+++ labpack/speech/aws/__init__.py
@@ -1,0 +1,4 @@
+''' a package of methods for handling aws speech recognition and synthesis apis '''
+__author__ = 'rcj1492'
+__created__ = '2017.11'
+__license__ = 'MIT'